#11: Rename the fakes + cleanup the documentations for release
