+++ src/main/java/io/github/dgroup/term4j/runtime/AppException.java
@@ -21,7 +21,7 @@
  * ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE
  * OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package io.github.dgroup.term4j;
+package io.github.dgroup.term4j.runtime;
 
 import org.cactoos.Text;
 import org.cactoos.iterable.IterableOf;
